+++ .gitattributes
@@ -1,4 +1,5 @@
 * text=auto
 test/statsCases/* eol=lf
 examples/* eol=lf
-eol=lf
+bin/* eol=lf
+*.svg eol=lf
