+++ test/configCases/url/file-loader/index.js
@@ -1,4 +1,4 @@
-import url from "../../../fixtures/images/file.png";
+import url from "../_images/file.png";
 
 it("should use file-loader", () => {
 	expect(url).toEqual("file-loader.png");
