+++ test/configCases/url/path/index.js
@@ -1,4 +1,4 @@
-import url from "../../../fixtures/images/file.png";
+import url from "../_images/file.png";
 
 it("should output asset with path", () => {
 	expect(url).toEqual("images/file.png");
