+++ test/configCases/url/publicPath/index.js
@@ -1,4 +1,4 @@
-import url from "../../../fixtures/images/file.png";
+import url from "../_images/file.png";
 
 it("should import asset with correct publicPath", () => {
 	expect(url).toEqual("assets/file.png");
