+++ test/configCases/url/types/index.js
@@ -1,9 +1,9 @@
-import png from "../../../fixtures/images/file.png";
-import svg from "../../../fixtures/images/file.svg";
-import jpg from "../../../fixtures/images/file.jpg";
+import png from "../_images/file.png";
+import svg from "../_images/file.svg";
+import jpg from "../_images/file.jpg";
 
 it("should output various asset types", () => {
-	expect(png).toEqual("990800716314551ab975.png");
-	expect(svg).toEqual("6003210f173dc7ede7a6.svg");
-	expect(jpg).toEqual("6c5cbdb49c76fee87d12.jpg");
+	expect(png).toEqual("83d6dab543c538e6621f.png");
+	expect(svg).toEqual("7c14c134bd6b6bedc352.svg");
+	expect(jpg).toEqual("00891fd33cfbdc56d145.jpg");
 });
